requirements.txt
# requirements.txt
# Основные зависимости для ARIA

# Ядро системы
numpy>=1.20.0
PyYAML>=6.0

# Для Python < 3.7 (если нужно)
dataclasses;python_version<'3.7'

# Опциональные зависимости (устанавливать при необходимости)
# pip install aria-ai[full] для полной установки

# Для обработки изображений
# Pillow>=9.0

# Для обработки аудио  
# librosa>=0.9.0

# Для обработки видео
# opencv-python>=4.5.0

# Для визуализации
# matplotlib>=3.5.0

# Для разработки
# pytest>=7.0
# pytest-cov>=4.0
# black>=22.0
# flake8>=5.0

# Для документации
# sphinx>=5.0
# sphinx-rtd-theme>=1.0